dashboards/crm_report/requirements.txt
flask>=2.3
gunicorn>=21.0
python-dotenv>=1.0
pandas>=1.5
paramiko==3.5.0
flask-limiter>=3.0
